requirements.txt
pyte==0.8.2
pywinpty==2.0.12
sv-ttk==2.6.0
wcwidth==0.2.13
